Move temporary files-related methods into TemporaryFilesUtil class
We use TestUtil in testing.

TestUtil has very low cohesion; it is just a giant mix of methods thrown together with little design direction. It would be great if the methods could be categorized and split up into more well-defined utility classes.

Let's move out all temporary files-related methods into TemporaryFilesUtil class.
